requirements.txt
asgiref==3.3.1
Django==3.1.5
django-debug-toolbar==3.2
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-jsonapi==4.0.0
inflection==0.5.1
psycopg2==2.8.6
pytz==2020.5
sqlparse==0.4.1
